requirements.txt
# Main requirements
Django==5.0
dj-database-url==0.2.2
django-extensions==3.2.1
django-filter==24.2
djangorestframework==3.15.1
geopy==2.4.1
Markdown==3.6
pandas==2.1.4
psycopg==3.1.17
tqdm==4.66.1

# Development requirements
black==23.7.0
django-silk==5.1.0
pylint==3.1.0
